leds: lp5521/5523: Fix multiple engine usage bug
Whenever the engine is loaded by the user-application, the operation mode is
reset first. But it has a problem in case of multiple engine used because
previous engine settings are cleared.
The driver should update not whole 8bits but each engine bit by masking.

On the other hands, whole engines should be reset when the driver is unloaded
and on initializing the LP5523 driver.
So, new functions are used for this handling - lp5521/5523_stop_all_engines().

Cc: Pali Rohár <pali.rohar@gmail.com>
Signed-off-by: Milo Kim <milo.kim@ti.com>
Signed-off-by: Bryan Wu <cooloney@gmail.com>

(cherry picked from commit 28c9266b38a00a07497daad0237f7fa154652ece)
[backport to 3.10: the hunk renaming the call in lp5523_init_program_engine
is dropped -- that function does not exist in this tree's older base]
Signed-off-by: arttttt <artem-bambalov@yandex.ru>