ext4: fix return value for ext4_es_scan()
Between 3.10 and 3.18, the abstraction to scan for objects in the slab
cache which can be freed when the system is under memory pressure
changed.  When I backported the ext4 code from 3.18 to the 3.10
kernel, I didn't get the return value required by the calling
conventions for the scan function correct, which could potentially
cause the memory reclaimer to loop indefinitely.

Bug=21907245

Change-Id: I3b13c9878663f409a100123ec3890114beb64172
Signed-off-by: "Theodore Ts'o" <tytso@google.com>