arch: arm64: boot: dts: HDMI Support
Remove HDMI node in PCL list of devices as
camera SW stack path is not supported for this driver.
IMX274 module cannot be supported as default board
and entries in plugin manager is causing the HDMI
driver to stay in disabled state.
Fix made to remove stray entries which disable HDMI
in DT for flash target IMX274 module.

Bug 200178647

Change-Id: I75f73c8177b088fa703fac986842d90323d8c981
Signed-off-by: Bhanu Murthy V <bmurthyv@nvidia.com>
Reviewed-on: http://git-master/r/1162034
GVS: Gerrit_Virtual_Submit
Reviewed-by: Shiva Dubey <sdubey@nvidia.com>
Reviewed-by: Hayden Du <haydend@nvidia.com>