Revert "HACK Revert "ARM: 7405/1: kexec: call platform_cpu_kill on the killer rather than the victim""
This reverts commit f207c657eabbc231bddfe3fa72a98de16eb2d2b0.

The HACK Revert moved the platform_cpu_kill() back to the "victim"
from the "killer". However, the platform_cpu_kill() is a function
calling tegra_cpu_kill() which is supposed to run on the "killer"
to poll the status of the "victim" via AFFINITY_INFO command.

Meanwhile, the tegra_cpu_kill calls msleep(), so putting it into
an IRQ context results in "BUG: scheduling while atomic".

So we revert this HACK since it doesn't break reboot any more.

Bug 200224278
Bug 1808023

Change-Id: Iaa63f1a848378df47d3f76821d641ed187983183
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1228810
GVS: Gerrit_Virtual_Submit
Tested-by: Vaibhav Shinde <vashinde@nvidia.com>
Reviewed-by: Vaibhav Shinde <vashinde@nvidia.com>
Reviewed-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-by: Dhiren Parmar <dparmar@nvidia.com>