devfreq: governor_pod: Raise threshold for vic03
The threshold used to govern the vic03 clock rate
is set at 250 (25% of the total capacity) which
seems to be too low as it's easier for devfreq to
drive the clock higher and then impact the power.

So this patch raises the threshold further to 30%
so as not to make the frequency fluctuate much.

Bug 200228585

Change-Id: Id13950bf170b21c8f6f9caf670efff25c4341f67
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1236174
GVS: Gerrit_Virtual_Submit
Reviewed-by: Somdutta Roy <somduttar@nvidia.com>
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>
Tested-by: Somdutta Roy <somduttar@nvidia.com>