video: tegra: hdmi2.0: fix hpd state in recovery
bug 200242122
bug 200243233

In kernel recovery path, we accidently disable hdmi
monitor in hpd event. This causes no display on hdmi
monitor in recovery mode which is 480p.
In normal scenario, rest of hpd engine takes care
but recovery, hpd is disabled causing this issue.
The hpd state is now checked which is set by bl for
this specific scenario.

Change-Id: I8812753927a2c6845c6b1c58cde83e91abbbbec0
Signed-off-by: Prafull Suryawanshi <prafulls@nvidia.com>
Reviewed-on: http://git-master/r/1241478
Reviewed-by: Aly Hirani <ahirani@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>