video: tegra: host: Optimize the accuracy of actmon average load
According to Bug 1381833, we can enlarge the k value and sample period
to increase the accuracy of the average load value from the activity
monitor. (The range of k is [1, 6] while the sample period should not
exceed 22us.)

Based on the experiments mentioned in the Bug 1828143, we found that
setting k=5 and sample period=20us could be relatively benifit to the
accuracy of the average load.

So this patch just modifies these two values.

Bug 1828143

Change-Id: I6fbd4d6a832248e121f177cc5e30640d3702e84c
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: http://git-master/r/1248444
GVS: Gerrit_Virtual_Submit
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>