Elevation of privilege vulnerability in kernel networking subsystem
An elevation of privilege vulnerability in the kernel networking
subsystem could enable a local malicious application to execute
arbitrary code within the context of the kernel. This issue is
rated as Moderate because it first requires compromising a
privileged process and current compiler optimizations restrict
access to the vulnerable code.

There is no validation of the len variable passed to the
ping_common_sendmsg function to check if it is less than
icmph_len leading to a potential overflow. The fix is designed
to add additional validation to prevent the potential overflow.

CVE-2016-8399
A-31349935
Bug 1836932

Change-Id: Ia61de145bd5e12c1f30847812abd06334054b416
Signed-off-by: Mithun Maragiri <mmaragiri@nvidia.com>
Signed-off-by: Gagan Grover <ggrover@nvidia.com>
Reviewed-on: http://git-master/r/1262344
Reviewed-by: Dhiren Parmar <dparmar@nvidia.com>