Denial of service vulnerability in kernel sound driver
A denial of service vulnerability in the kernel could allow a
local malicious application to cause a device reboot.
This issue is rated as Low because it is a temporary denial of
service.

The original fix used -EIO as the error return code but
the function signatures had unsigned int as the return type.
The updated fix uses -1 as the error return code instead of -EIO
so the error return code is more clearly defined.

CVE-2016-6690
A-28838221
Bug 1836932

Change-Id: I10754b638b7432242d7baa1355d35bf56c2ad085
Signed-off-by: Mithun Maragiri <mmaragiri@nvidia.com>
Signed-off-by: Gagan Grover <ggrover@nvidia.com>
Reviewed-on: http://git-master/r/1262338
Reviewed-by: Dhiren Parmar <dparmar@nvidia.com>