media: usb: em28xx: resolve WinTV dualHD plugging/unplugging crash
The second adapter has to be removed first because the data structure of
the first adapter contains a pointer to the second. If memory of the
first adapter is freed before the second adapter is removed the pointer
could be corrupted and kernel crashes could follow.
SHA1 commit:d86686b5879b5fba57a0fac47d94982cafaa65bb

Bug: 32669837
Bug 1904252

Change-Id: I691955e8d4321a18d8850c9662948005b6614465
Signed-off-by: Jean Huang <jeanh@nvidia.com>
Reviewed-on: http://git-master/r/1468915
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Manish Tuteja <mtuteja@nvidia.com>