tegra:dc: reset background color for YUV420/10
Reset background color for YUV420 10-bit packing to black. Black
background is necessary for the optimized packing scheme. The clear
layer needs to be presented only if no layer is otherwise visible.

The scheme combines two layers with BLEND_ADD blending mode. The
layers are displayed with one pixel horizontal offset, hence left
and right screen edge is covered by only one of the layers. That
layer is combined with the background. Therefore the background
must be black in order not to interfere with the layer content.

Bug 200318984

Change-Id: Ic17b224bdbff3185bf393d70a6d0975e538f75d9
Signed-off-by: Michael Frydrych <mfrydrych@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1512745
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Season Li <seasonl@nvidia.com>
Tested-by: Season Li <seasonl@nvidia.com>
Reviewed-by: Ivan Raul Guadarrama <iguadarrama@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>